+++ src/protobuf-net/Meta/MetaType.cs
@@ -1250,6 +1250,8 @@ namespace ProtoBuf.Meta
                     case ProtoTypeCode.UInt64: defaultValue = (ulong)0; break;
                     case ProtoTypeCode.TimeSpan: defaultValue = TimeSpan.Zero; break;
                     case ProtoTypeCode.Guid: defaultValue = Guid.Empty; break;
+                    case ProtoTypeCode.IntPtr: defaultValue = IntPtr.Zero; break;
+                    case ProtoTypeCode.UIntPtr: defaultValue = UIntPtr.Zero; break;
                 }
             }
             if ((attrib = GetAttribute(attribs, "System.ComponentModel.DefaultValueAttribute")) is not null)
@@ -2367,6 +2369,8 @@ namespace ProtoBuf.Meta
                     case ProtoTypeCode.UInt16: return ((ushort)value) == 0;
                     case ProtoTypeCode.UInt32: return ((uint)value) == 0;
                     case ProtoTypeCode.UInt64: return ((ulong)value) == 0;
+                    case ProtoTypeCode.IntPtr: return ((IntPtr)value) == IntPtr.Zero;
+                    case ProtoTypeCode.UIntPtr: return ((UIntPtr)value) == UIntPtr.Zero;
                 }
             }
             catch { }